factor out roomGenerator
